requirements.txt
fastapi
uvicorn
numpy
opencv-python
spacy
Pillow
easyocr
python-dotenv
python-multipart
torch
torchvision
